Time: 2591 ms (5.02%), Space: 181.3 MB (32.36%) - LeetHub

+++ 3013-divide-an-array-into-subarrays-with-minimum-cost-ii/3013-divide-an-array-into-subarrays-with-minimum-cost-ii.cpp
@@ -3,16 +3,16 @@ public:
     typedef long long ll;
     long long minimumCost(vector<int>& v, int k, int dist) {
         k -= 1;
-        multiset<ll> right;
-        multiset<ll, greater<ll>> left;
-        ll start = 1LL, end = 1LL, cnt = 0LL, ans = 1e18, n = v.size();
+        multiset<int> right;
+        multiset<int, greater<int>> left;
+        ll start = 1, end = 1, cnt = 0, ans = 1e18, n = v.size();
         while(end < n) {
-            cnt += (ll)v[end];
+            cnt += v[end];
             left.insert(v[end]);
             end += 1;
             if(!(left.size() || right.size()) && *left.begin() > *right.begin()) {
-                ll a = *left.begin();
-                ll b = *right.begin();
+                int a = *left.begin();
+                int b = *right.begin();
                 cnt += (b - a);
                 left.erase(left.begin());
                 right.erase(right.begin());
@@ -20,7 +20,7 @@ public:
                 left.insert(b);
             }
             if(left.size() > k) {
-                ll a = *left.begin();
+                int a = *left.begin();
                 cnt -= a;
                 right.insert(a);
                 left.erase(left.begin());
@@ -46,6 +46,7 @@ public:
             }
             
         }
+        
         return ans + v[0];
     }
 };
